backend/requirements.txt
# ML and Data Processing
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
xgboost==2.0.3
joblib==1.3.2

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
supabase==2.3.0

# HTTP Requests
requests==2.31.0
httpx


# Environment
python-dotenv==1.0.0

# AWS SDK
boto3==1.29.7

# Data Collection
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dateutil==2.8.2

